+++ CMS/admin/partials/view_all_comments.php
@@ -1,3 +1,4 @@
+
 					<table class="table table-bordered table-hover">
 						<thead>
 							<tr>
@@ -68,16 +69,23 @@
 	<?php
 
 		if(isset($_GET['delete'])) {
-			$delete_comment_id = $_GET['delete'];
-			$query = "DELETE FROM comments WHERE comment_id = {$delete_comment_id} ";
-			$delete_query = mysqli_query($connection, $query);
-			confirm($delete_query);
-			header("Location: comments.php");
+
+			if(isset($_SESSION['user_role'])) {
+				if($_SESSION['user_role'] == 'admin') {
+					$delete_comment_id = escape($_GET['delete']);
+					$query = "DELETE FROM comments WHERE comment_id = {$delete_comment_id} ";
+					$delete_query = mysqli_query($connection, $query);
+					confirm($delete_query);
+					header("Location: comments.php");
+				}
+			}
+
+
 		}
 //							Set approved and unapproved in comments table
 
 			if(isset($_GET['unapproved'])) {
-			$the_comment_id = $_GET['unapproved'];
+			$the_comment_id = escape($_GET['unapproved']);
 			$query = "UPDATE comments SET comment_status = 'unapproved' WHERE comment_id = $the_comment_id ";
 			$unapproved_comment_query = mysqli_query($connection, $query);
 			confirm($unapproved_comment_query);
@@ -85,7 +93,7 @@
 			}
 
 			if(isset($_GET['approved'])) {
-			$the_comment_id = $_GET['approved'];
+			$the_comment_id = escape($_GET['approved']);
 			$query = "UPDATE comments SET comment_status = 'approved' WHERE comment_id = $the_comment_id ";
 			$approved_comment_query = mysqli_query($connection, $query);
 			confirm($approved_comment_query);
